# contracts/openapi.json
{
  "components": {
    "responses": {
      "ErrorResponse": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Stable machine-readable error envelope."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "type": "object"
      },
      "ExpenseBeneficiaryResponse": {
        "additionalProperties": false,
        "description": "Beneficiary history resolved to current participant metadata.",
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "participant_id": {
            "title": "Participant Id",
            "type": "string"
          }
        },
        "required": [
          "participant_id",
          "name",
          "archived"
        ],
        "title": "ExpenseBeneficiaryResponse",
        "type": "object"
      },
      "ExpenseContributorRequest": {
        "additionalProperties": false,
        "description": "One contributor and its decimal lexical input.",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "participant_id": {
            "title": "Participant Id",
            "type": "string"
          }
        },
        "required": [
          "participant_id",
          "amount"
        ],
        "title": "ExpenseContributorRequest",
        "type": "object"
      },
      "ExpenseContributorResponse": {
        "additionalProperties": false,
        "description": "Contributor history resolved to current participant metadata.",
        "properties": {
          "amount_cents": {
            "title": "Amount Cents",
            "type": "integer"
          },
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "participant_id": {
            "title": "Participant Id",
            "type": "string"
          }
        },
        "required": [
          "participant_id",
          "name",
          "archived",
          "amount_cents"
        ],
        "title": "ExpenseContributorResponse",
        "type": "object"
      },
      "ExpenseResponse": {
        "additionalProperties": false,
        "description": "Persisted source expense; monetary values are integer cents only.",
        "properties": {
          "amount_cents": {
            "title": "Amount Cents",
            "type": "integer"
          },
          "beneficiaries": {
            "items": {
              "$ref": "#/components/schemas/ExpenseBeneficiaryResponse"
            },
            "title": "Beneficiaries",
            "type": "array"
          },
          "contributors": {
            "items": {
              "$ref": "#/components/schemas/ExpenseContributorResponse"
            },
            "title": "Contributors",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "group_id": {
            "title": "Group Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outing Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "group_id",
          "description",
          "amount_cents",
          "contributors",
          "beneficiaries"
        ],
        "title": "ExpenseResponse",
        "type": "object"
      },
      "ExpenseWriteRequest": {
        "additionalProperties": false,
        "description": "Complete replacement command for an expense.",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "beneficiary_ids": {
            "items": {
              "type": "string"
            },
            "title": "Beneficiary Ids",
            "type": "array"
          },
          "contributors": {
            "items": {
              "$ref": "#/components/schemas/ExpenseContributorRequest"
            },
            "title": "Contributors",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "outing_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outing Id"
          }
        },
        "required": [
          "description",
          "amount",
          "contributors",
          "beneficiary_ids"
        ],
        "title": "ExpenseWriteRequest",
        "type": "object"
      },
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cc_session": {
        "description": "Opaque server-recognized session cookie.",
        "in": "cookie",
        "name": "cc_session",
        "type": "apiKey"
      },
      "csrf_token": {
        "description": "Required on unsafe cookie-authenticated operations.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Cuentas Claras API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/groups/{group_id}/expenses": {
      "get": {
        "description": "List source expenses in stable creation order with current names.",
        "operationId": "list_expenses_api_v1_groups__group_id__expenses_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExpenseResponse"
                  },
                  "title": "Response List Expenses Api V1 Groups  Group Id  Expenses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "List Expenses",
        "tags": [
          "expenses"
        ]
      },
      "post": {
        "description": "Parse lexical money and create one complete source expense.",
        "operationId": "create_expense_api_v1_groups__group_id__expenses_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Create Expense",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/v1/groups/{group_id}/expenses/{expense_id}": {
      "delete": {
        "description": "Delete a source expense and all of its derived effect.",
        "operationId": "delete_expense_api_v1_groups__group_id__expenses__expense_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "title": "Expense Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Delete Expense",
        "tags": [
          "expenses"
        ]
      },
      "get": {
        "description": "Read one group-owned source expense.",
        "operationId": "get_expense_api_v1_groups__group_id__expenses__expense_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "title": "Expense Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Get Expense",
        "tags": [
          "expenses"
        ]
      },
      "patch": {
        "description": "Validate a full replacement before changing the source expense.",
        "operationId": "edit_expense_api_v1_groups__group_id__expenses__expense_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "title": "Expense Id",
              "type": "string"
            }
          },
          {
            "description": "Must match the readable cc_csrf cookie.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "cc_session": []
          }
        ],
        "summary": "Edit Expense",
        "tags": [
          "expenses"
        ]
      }
    },
...
  }
}